+++ perseo-quality/src/perseo_quality/radiometric_analysis/custom_dataclasses.py
@@ -136,7 +136,7 @@ class ScallopingRadiometricKPI(BaseRadiometricKPI):
 class RadiometricOutputProductGeneralInfo:
     """Block-wise Radiometric analysis product related generic info"""
 
-    product_name: str
+    product: str
     channel: str
     swath: str
     polarization: str