+++ perseo-quality/tests/unit_tests/radiometric_analysis/test_radiometric_analysis_plot.py
@@ -32,7 +32,7 @@ class Radiometric2DHistPlotTest(unittest.TestCase):
         profiles[2, :] *= 125
         data = RadiometricProfilesOutput(
             general_info=RadiometricOutputProductGeneralInfo(
-                product_name="test",
+                product="test",
                 channel="1",
                 swath="S1",
                 acquisition_mode="SCANSAR",